requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pathlib==1.0.1
watchdog==3.0.0
cssutils
fastapi
uvicorn
psutil
email_validator
cssutils
python-jose
passlib